[cli] Rename `account submit-multisig-txn` command to `account submit-txn`
